categories completed

+++ src/main/java/com/ishwor/newp/spring/boot/repository/categories/CategoriesRepository.java
@@ -7,10 +7,10 @@ import com.ishwor.newp.spring.boot.domain.Categories;
 public interface CategoriesRepository {
 	public List<Categories> findAllCategories();
 
-	public Categories findBycategoriesId(short id);
+	public Categories findBycategoriesId(int id);
 
 	public void addCategories(Categories categories);
 
-	public void removeCategories(short id);
+	public void removeCategories(int id);
 
 }
