+++ src/main/java/com/ishwor/newp/spring/boot/service/categories/CategoriesService.java
@@ -7,9 +7,9 @@ import com.ishwor.newp.spring.boot.domain.Categories;
 public interface CategoriesService {
 	public List<Categories> findAllCategories();
 
-	public Categories findBycategoriesId(short id);
+	public Categories findBycategoriesId(int id);
 
 	public void addCategories(Categories categories);
 
-	public void removeCategories(short id);
+	public void removeCategories(int id);
 }
